refactor(desktop): split agents page package boundaries
